api/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
python-dotenv
alembic
bcrypt>=4.0.0
passlib[bcrypt]
python-jose
pydantic_settings
python-multipart
cryptography